requirements.txt
py-solc-x==2.0.3
argparse==1.4.0
# web3==7.3.1 (pos)
# web3==5.31.1 (pow)
web3==5.31.1
